Add Windows support for toolbox image
We add Windows images to the antrea/toolbox manifest (for both the 1809
and ltsc2022 Windows nanoserver versions).

The Windows image is based directly on K8s pause image. It includes
wincat, pause, and of course everything included by default in the
nanoserver image (such as ping).

In the future, we may want to install additional utilities in the
Windows image.

Because we share the same versioning between Linux and Windows, the
toolbox version is bumped up to 1.3.

Fixes #26

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>